+++ samcarter/whitewaterride/convert2video.sh
@@ -15,6 +15,8 @@ ffmpeg  \
 HandBrakeCLI --crop 0:0:0:0  -i whitewaterride_raw.mp4 -o whitewaterride.mp4
 
 # clean up
-rm whitewaterride-*.png
-rm water-*.jpg
-rm whitewaterride_raw.mp4
+if [ -e "whitewaterride.mp4" ]; then
+  rm whitewaterride-*.png
+  rm water-*.jpg
+  rm whitewaterride_raw.mp4
+fi
